+++ src/main/java/Variable.java
@@ -1,11 +1,8 @@
-/*
- * Created 24-Jul-2024
- */
 
 public class Variable {
 
     public String name;
-    public float value;
+    public double value;
 
     public Variable() {
         name = "";
